% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{area_to_short}
\alias{area_to_short}
\title{Convert Area Number Vector to Short Name Factor}
\usage{
area_to_short(n)
}
\arguments{
\item{n}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=factor]{factor()}}
}
\description{
Convert Area Number Vector to Short Name Factor
}
