% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{fit_movement_model}
\alias{fit_movement_model}
\title{Fit Movement Model}
\usage{
fit_movement_model(
  spatial_name,
  temporal_name,
  grouping_name,
  liberty_limited = FALSE,
  chains,
  nuts_step_size,
  iter_warmup,
  iter_sampling,
  use_reduce_sum,
  threads_per_chain,
  tags_released,
  tags_recovered,
  harvest_rates,
  released_time_unit,
  released_time_max,
  liberty_time_max,
  colname_released_date,
  colname_released_area,
  colname_group,
  colname_recovered_date,
  colname_recovered_area,
  colname_id,
  area_list,
  group_list,
  released_date_start,
  released_date_end,
  movement_time_max,
  harvest_time_max,
  reporting_time_max,
  tag_loss_rate_initial,
  tag_loss_rate_ongoing,
  natural_mortality_rate,
  reporting_rates,
  random_walk,
  movement_rate_fudge = 1e-12,
  predicted_tags_fudge = 1e-12,
  h_prior_mean = NULL,
  h_prior_sd,
  phi_prior_mean,
  phi_prior_sd,
  sigma_prior_mean = numeric(0),
  sigma_prior_sd = numeric(0),
  ...
)
}
\arguments{
\item{spatial_name}{\code{\link[=character]{character()}}}

\item{temporal_name}{\code{\link[=character]{character()}}}

\item{grouping_name}{\code{\link[=character]{character()}}}

\item{liberty_limited}{\code{\link[=logical]{logical()}}}

\item{chains}{\code{\link[=numeric]{numeric()}}}

\item{nuts_step_size}{\code{\link[=numeric]{numeric()}}}

\item{iter_warmup}{\code{\link[=numeric]{numeric()}}}

\item{iter_sampling}{\code{\link[=numeric]{numeric()}}}

\item{use_reduce_sum}{\code{\link[=logical]{logical()}}}

\item{threads_per_chain}{\code{\link[=numeric]{numeric()}}}

\item{tags_released}{\code{\link[=data.frame]{data.frame()}}}

\item{tags_recovered}{\code{\link[=data.frame]{data.frame()}}}

\item{harvest_rates}{\code{\link[=data.frame]{data.frame()}}}

\item{released_time_unit}{\code{\link[=character]{character()}}}

\item{released_time_max}{\code{\link[=numeric]{numeric()}}}

\item{liberty_time_max}{\code{\link[=numeric]{numeric()}}}

\item{colname_released_date}{\code{\link[=character]{character()}}}

\item{colname_released_area}{\code{\link[=character]{character()}}}

\item{colname_group}{\code{\link[=character]{character()}}}

\item{colname_recovered_date}{\code{\link[=character]{character()}}}

\item{colname_recovered_area}{\code{\link[=character]{character()}}}

\item{colname_id}{\code{\link[=character]{character()}}}

\item{area_list}{\code{\link[=list]{list()}}}

\item{group_list}{\code{\link[=list]{list()}}}

\item{released_date_start}{\code{\link[=character]{character()}}}

\item{released_date_end}{\code{\link[=character]{character()}}}

\item{movement_time_max}{\code{\link[=numeric]{numeric()}}}

\item{harvest_time_max}{\code{\link[=numeric]{numeric()}}}

\item{reporting_time_max}{\code{\link[=numeric]{numeric()}}}

\item{tag_loss_rate_initial}{\code{\link[=numeric]{numeric()}}}

\item{tag_loss_rate_ongoing}{\code{\link[=numeric]{numeric()}}}

\item{natural_mortality_rate}{\code{\link[=numeric]{numeric()}}}

\item{reporting_rates}{\code{\link[=numeric]{numeric()}}}

\item{random_walk}{\code{\link[=numeric]{numeric()}}}

\item{movement_rate_fudge}{\code{\link[=numeric]{numeric()}}}

\item{predicted_tags_fudge}{\code{\link[=numeric]{numeric()}}}

\item{h_prior_mean}{\code{\link[=numeric]{numeric()}}}

\item{h_prior_sd}{\code{\link[=numeric]{numeric()}}}

\item{phi_prior_mean}{\code{\link[=numeric]{numeric()}}}

\item{phi_prior_sd}{\code{\link[=numeric]{numeric()}}}

\item{sigma_prior_mean}{\code{\link[=numeric]{numeric()}}}

\item{sigma_prior_sd}{\code{\link[=numeric]{numeric()}}}

\item{...}{additional arguments to pass to cmdstanr \code{$sample()} method}
}
\value{
\code{\link[mmmstan:mmmfit]{mmmstan::mmmfit()}} object
}
\description{
Fit Movement Model
}
