% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fits}
\alias{fits}
\alias{fit_region_average_pooled}
\title{Sablefish Movement Model Fit Parameter Summaries}
\format{
An object of class \code{list} of length 4.
}
\usage{
fit_region_average_pooled
}
\description{
Sablefish Movement Model Fit Parameter Summaries
}
\keyword{datasets}
