% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harvest_rates}
\alias{harvest_rates}
\title{Sablefish Harvest Rates}
\format{
\code{\link[=data.frame]{data.frame()}} \code{\link[=tibble]{tibble()}} with 760 rows and 6 variables:
\describe{
\item{spatial}{spatial scale, one of \code{region},
\code{subregion}, or \code{omregion}}
\item{name}{spatial strata name}
\item{short}{short version of spatial strata name}
\item{number}{spatial strata number}
\item{year}{year}
\item{harvest_rate}{sablefish harvest rate}
}
}
\source{
Harvest rate data come from:
\describe{
\item{regions}{
\itemize{
\item{Alaska: sablefish stock assessment annual fishing mortality
rate F, converted to a harvest rate by \code{H = 1 - exp(-F)}}
\item{British Columbia: sablefish stock assessment annual legal
harvest rate}
\item{California Current: sablefish stock assessment annual fishing
mortality rate F, converted to a harvest rate by
\code{H = 1 - exp(-F)}}
}
}
\item{subregions}{
\itemize{
\item{Southeast Alaska: Chatham Strait sablefish statistical
catch-at-age model annual fishing mortality rate F, converted to a
harvest rate by \code{H = 1 - exp(-F)}}
\item{Other Alaska subregions: sablefish stock assessment annual catch
(including discards) by subregion divided by annual age 4+ biomass
by subregion}
\item{British Columbia subregions: same as British Columbia}
\item{California Current subregions: same as California Current}
}
}
\item{omregions}{
\itemize{
\item{Western Alaska: same as Alaska}
\item{Eastern Alaska: same as Southeast Alaska}
\item{British Columbia operating model strata:
same as British Columbia}
\item{California Current operating model strata: same as
California Current}
}
}
}

see https://github.com/luke-a-rogers/sablefish-data
}
\usage{
harvest_rates
}
\description{
A dataset containing annual sablefish harvest rates for spatial strata
in the northeast Pacific Ocean.
}
\keyword{datasets}
