% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heat_map_matrix}
\alias{heat_map_matrix}
\title{Heat Map Matrix}
\usage{
heat_map_matrix(
  data,
  movement_time = 1,
  released_group = 1,
  xlab = NULL,
  ylab = NULL,
  xtext = TRUE,
  ytext = TRUE,
  margin_x = -15,
  margin_y = -15,
  font_size_mean = 3,
  font_nudge_mean = 0.15,
  font_size_sd = 2,
  font_nudge_sd = 0.15,
  legend_name = "Movement rate"
)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{movement_time}{\code{\link[=numeric]{numeric()}}}

\item{released_group}{\code{\link[=numeric]{numeric()}}}

\item{xlab}{\code{\link[=character]{character()}} or \code{NULL}}

\item{ylab}{\code{\link[=character]{character()}} or \code{NULL}}

\item{legend_name}{\code{\link[=character]{character()}} or \code{NULL}}

\item{font_size_p}{\code{\link[=numeric]{numeric()}}}

\item{font_nudge_p}{\code{\link[=numeric]{numeric()}}}

\item{font_size_ci}{\code{\link[=numeric]{numeric()}}}

\item{font_nudge_ci}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Heat Map Matrix
}
