% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mod}
\alias{mod}
\title{Modulo Function}
\usage{
mod(a, b, zeros = TRUE)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{b}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{zeros}{\code{\link[=logical]{logical()}} replace zeros by \code{b}?}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Modulo Function
}
\examples{
# Standard modulo operation
mod(1:10, 8)

# Replace zeros by \code{b}
mod(1:10, 8, zeros = FALSE)

}
