% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numbers_at_length}
\alias{numbers_at_length}
\title{Sablefish Numbers At Length}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3720 rows and 6 columns.
}
\source{
see https://github.com/luke-a-rogers/sablefish-data
}
\usage{
numbers_at_length
}
\description{
Sablefish Numbers At Length
}
\keyword{datasets}
