% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heat_composite}
\alias{plot_heat_composite}
\title{Plot Heat Map Composite Figure}
\usage{
plot_heat_composite(data1, data2)
}
\arguments{
\item{data1}{\code{\link[=data.frame]{data.frame()}}}

\item{data2}{\code{\link[=data.frame]{data.frame()}}}
}
\value{
\code{\link[=character]{character()}} file path
}
\description{
Plot Heat Map Composite Figure
}
