% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{point_movement}
\alias{point_movement}
\title{Point Movement Plot}
\usage{
point_movement(
  data,
  released_group = 1,
  point_size = 0.8,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{released_group}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Point Movement Plot
}
