% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pres_abundance_exchange}
\alias{pres_abundance_exchange}
\title{Presentation Abundance Exchange}
\usage{
pres_abundance_exchange(data, numbers, name, years, xlim, ylim, width, height)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{numbers}{\code{\link[=data.frame]{data.frame()}}}

\item{name}{\code{\link[=character]{character()}}}

\item{years}{\code{\link[=numeric]{numeric()}}}

\item{xlim}{\code{\link[=numeric]{numeric()}}}

\item{ylim}{\code{\link[=numeric]{numeric()}}}

\item{width}{\code{\link[=numeric]{numeric()}}}

\item{height}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=character]{character()}} file path
}
\description{
Presentation Abundance Exchange
}
