% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pres_bar_quarter}
\alias{pres_bar_quarter}
\title{Presentation Bar Quarter}
\usage{
pres_bar_quarter(
  data,
  name,
  released_group = 1,
  xlab = "Quarter",
  ylab = "Movement rate",
  ylim = c(0, 1),
  xtext = TRUE,
  ytext = TRUE,
  width = 5,
  height = 3,
  legend_name = "Region"
)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{name}{\code{\link[=character]{character()}}}

\item{released_group}{\code{\link[=numeric]{numeric()}}}

\item{xlab}{\code{\link[=character]{character()}}}

\item{ylab}{\code{\link[=character]{character()}}}

\item{ylim}{\code{\link[=numeric]{numeric()}}}

\item{xtext}{\code{\link[=logical]{logical()}}}

\item{ytext}{\code{\link[=logical]{logical()}}}

\item{width}{\code{\link[=numeric]{numeric()}}}

\item{height}{\code{\link[=numeric]{numeric()}}}

\item{legend_name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[=character]{character()}} file path
}
\description{
Presentation Bar Quarter
}
