% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pres_harvest_length}
\alias{pres_harvest_length}
\title{Presentation Harvest Rates By Length}
\usage{
pres_harvest_length(
  data,
  name,
  length_class = c("Small", "Large"),
  harvest_group = c(1, 2),
  xlab = "Year",
  ylab = "Harvest rate",
  xlim = c(1979, 2020),
  ylim = c(0, 0.2),
  point_size = 0.8,
  width = 5,
  height = 4
)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{name}{\code{\link[=character]{character()}}}

\item{harvest_group}{\code{\link[=numeric]{numeric()}}}

\item{xlab}{\code{\link[=character]{character()}}}

\item{ylab}{\code{\link[=character]{character()}}}

\item{xlim}{\code{\link[=numeric]{numeric()}}}

\item{ylim}{\code{\link[=numeric]{numeric()}}}

\item{point_size}{\code{\link[=numeric]{numeric()}}}

\item{width}{\code{\link[=numeric]{numeric()}}}

\item{height}{\code{\link[=numeric]{numeric()}}}
}
\value{

}
\description{
Presentation Harvest Rates By Length
}
