% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pres_point_movement}
\alias{pres_point_movement}
\title{Presentation Point Movement Plot}
\usage{
pres_point_movement(
  data,
  name,
  released_group = 1,
  point_size = 0.8,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  width = 4,
  height = 4
)
}
\arguments{
\item{data}{\code{\link[=data.frame]{data.frame()}}}

\item{released_group}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Presentation Point Movement Plot
}
