% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_to_character}
\alias{round_to_character}
\title{Round to Character}
\usage{
round_to_character(x, digits)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} Value to round}

\item{digits}{\link{numeric} Number of digits}
}
\value{
\code{\link[=character]{character()}}
}
\description{
Round to Character
}
