% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tags_released}
\alias{tags_released}
\title{Sablefish Tags Released}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 944820 rows and 7 columns.
}
\source{
see https://github.com/luke-a-rogers/sablefish-data
}
\usage{
tags_released
}
\description{
A dataset containing sablefish tags released by date, released length,
and spatial stratum in the northeast Pacific Ocean.
}
\details{
\describe{
\item{tag_id}{}
\item{released_date}{}
\item{released_length}{}
\item{released_region}{}
\item{released_subregion}{}
\item{released_omregion}{}
\item{source}{}
}
}
\keyword{datasets}
