% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_to_label}
\alias{time_to_label}
\title{Time To Label}
\usage{
time_to_label(n)
}
\arguments{
\item{n}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=vector]{vector()}}
}
\description{
Time To Label
}
