% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unit_to_time}
\alias{unit_to_time}
\title{Convert Released Time Unit To Time Steps Per Year}
\usage{
unit_to_time(unit)
}
\arguments{
\item{unit}{\code{\link[=character]{character()}} one of \code{'year'}, \code{'quarter'}, or
\code{'month'}.}
}
\value{
\code{\link[=numeric]{numeric()}} number of time steps per year
}
\description{
Convert Released Time Unit To Time Steps Per Year
}
\examples{
unit_to_time("year")
unit_to_time("quarter")
unit_to_time("month")

}
