% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_data}
\alias{write_data}
\title{Write An Object To Data And Return The Path}
\usage{
write_data(...)
}
\arguments{
\item{...}{Unquoted name of an existing object to write to \code{data/}.
See \code{?usethis::use_data()}.}
}
\value{
\code{\link[=character]{character()}} file path
}
\description{
Write An Object To Data And Return The Path
}
